Moving `json_saving` to a more appropriate location
Now, there are fewer calls across components.
